requirements.txt
pipenv>=2022.5.2
django>=3.2.13
djangorestframework
django-cors-headers
django-admin
django-material
django-viewflow